로직수정
