--- src/App.css
@@ -1,4 +1,0 @@
-#root {
-  width: 100%;
-  text-align: center;
-}
