+++ src/GameBoard.tsx
@@ -1,12 +1,15 @@
-import React from "react";
 import styled from "@emotion/styled";
+import { useRef, useState, useEffect, useContext } from "react";
+import UserContext from "./userContext";
+import Modal from "./Modal";
 
-interface GameBoardProps {
-  row: number;
-  index: number;
-  numbers1: (number | null)[];
-  clickCount: number[];
-  buttonHandler: (idx: number) => void;
+interface GameButtonProps {
+  clickCount: number;
+}
+interface GameData {
+  time: string; // 날짜 문자열 (toString() 결과)
+  score: number; // 점수 (seconds 변수)
+  level: number; // 레벨 (row - 2 계산 결과)
 }
 const GameTable = styled.table`
   border-collapse: separate;
@@ -16,28 +19,188 @@ const GameTable = styled.table`
   height: 700px;
   text-align: center;
   table-layout: fixed;
-`
-const GameButton = styled.button`
-  flex:1;
-  border: 1px solid black; /* 각 셀 테두리 */
+`;
+
+const GameButton = styled.button<GameButtonProps>`
+  background-color: ${(props) => {
+    if (props.clickCount === 1) {
+      return "lightblue";
+    }
+    if (props.clickCount === 2) {
+      return "rgb(255, 255, 255)";
+    }
+    return "rgb(188, 188, 245)";
+  }};
+  flex: 1;
+  border: 1px solid black;
   text-align: center;
   padding: 0;
-  margin:0;
+  margin: 0;
   width: 100%;
   height: 100%;
   transition: all 0.1s ease-in-out;
-  :active {
-  background-color: aliceblue; /* 투명도를 줘서 깜빡이는 느낌 */
-}
-`
+  cursor: pointer;
 
-const GameBoard: React.FC<GameBoardProps> = ({
-  row,
-  index,
-  numbers1,
-  clickCount,
-  buttonHandler,
-}) => {
+  &:active {
+    background-color: aliceblue;
+  }
+
+  &:hover {
+    opacity: 0.8;
+  }
+`;
+
+const GameBoard = ({}) => {
+  const { level, setResults, seconds, setSeconds } = useContext(UserContext);
+  const row = level === "Lv1" ? 3 : level === "Lv2" ? 4 : 5;
+  const getRandomNumbers = (row: number): [number[], number[]] => {
+    const count = row * row;
+    const arr1 = Array.from({ length: count }, (_, i) => i + 1);
+    const arr2 = Array.from({ length: count }, (_, i) => i + 1 + count);
+    return [shuffleArray(arr1), shuffleArray(arr2)];
+  };
+  const shuffleArray = (arr: number[]): number[] => {
+    const shuffled = [...arr];
+    for (let i = shuffled.length - 1; i > 0; i--) {
+      const j = Math.floor(Math.random() * (i + 1));
+      [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]];
+    }
+    return shuffled;
+  };
+  const [date, setDate] = useState<string>("");
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const closeModal = () => setIsModalOpen(false);
+  const openModal = () => setIsModalOpen(true);
+  const [numbers1, setNumbers1] = useState<(number | null)[]>([]);
+  const [numbers2, setNumbers2] = useState<number[]>([]);
+  const [index, setIndex] = useState(0);
+  const [clickCount, setClickCount] = useState<number[]>(
+    Array(row * row).fill(0)
+  );
+  const intervalRef = useRef<number | null>(null); //현재 게임 시간
+  function clickCountPlus(idx: number) {
+    const prev = clickCount;
+    const newCounts = [...prev];
+    newCounts[idx] = prev[idx] + 1;
+    setClickCount(newCounts);
+  }
+  const startTimer = () => {
+    if (intervalRef.current !== null) return;
+
+    const startTime = Date.now();
+    setSeconds(0);
+
+    intervalRef.current = window.setInterval(() => {
+      const elapsed = (Date.now() - startTime) / 1000;
+      setSeconds(elapsed);
+    }, 10);
+  };
+  const stopTimer = () => {
+    if (intervalRef.current !== null) {
+      clearInterval(intervalRef.current);
+      intervalRef.current = null;
+    }
+    setIndex(0);
+    setSeconds(0);
+  };
+  useEffect(() => {
+    const [arr1, arr2] = getRandomNumbers(row);
+    setNumbers1(arr1);
+    setNumbers2(arr2);
+    setClickCount(Array(row * row).fill(0));
+    return () => stopTimer();
+  }, [level]);
+
+  const buttonHandler = (idx: number) => {
+    
+    if (index == 0 && numbers1[idx] == index + 1) {
+      // 게임 진행중이 아님
+      setIndex((prevIndex) => prevIndex + 1);
+      startTimer();
+      const newNumbers = [...numbers1]; // 기존 배열 복사
+      newNumbers[idx] = numbers2[idx]; // 수정
+      setNumbers1(newNumbers); // 상태 업데이트
+      clickCountPlus(idx);
+    }
+
+    // 게임 진행중이 맞음
+    else if (numbers1[idx] === row * row * 2) {
+      // 종료조건 먼저 검사
+      // 종료조건
+      stopTimer();
+      const dateNum = new Date(Date.now());
+      const formatted = dateNum.toLocaleString("ko-KR", {
+        year: "numeric",
+        month: "2-digit",
+        day: "2-digit",
+        hour: "2-digit",
+        minute: "2-digit",
+        hour12: false,
+      });
+      setDate(formatted);
+      const gameData = {
+        time: date.toString(),
+        score: seconds,
+        level: row - 2,
+      };
+      saveData(gameData);
+
+      openModal();
+      //alert(`점수: ${gameData.score}, 레벨: ${gameData.Lv}`);
+      const [arr1, arr2] = getRandomNumbers(row);
+      setNumbers1(arr1);
+      setNumbers2(arr2);
+      setClickCount(Array(row * row).fill(0));
+      setResults(JSON.parse(localStorage.getItem("gameResults") || "[]"));
+    } else if (numbers1[idx] === index + 1) {
+      if (numbers1[idx] != numbers2[idx]) {
+        const newNumbers = [...numbers1]; // 기존 배열 복사
+        newNumbers[idx] = numbers2[idx]; // 수정
+        setNumbers1(newNumbers); // 상태 업데이트
+
+        clickCountPlus(idx);
+      } else if (numbers1[idx] === numbers2[idx]) {
+        const newNumbers = [...numbers1]; // 기존 배열 복사
+        newNumbers[idx] = null; // 수정
+        setNumbers1(newNumbers);
+
+        clickCountPlus(idx);
+      }
+      setIndex((prevIndex) => prevIndex + 1);
+    }
+  };
+
+  const loadData = () => {
+    const latest = JSON.parse(localStorage.getItem("gameResults") || "[]");
+
+    // 최신 순으로 정렬
+    latest.sort(
+      (a: GameData, b: GameData) =>
+        new Date(b.time).getTime() - new Date(a.time).getTime()
+    );
+    return latest[0];
+  };
+
+  function saveData(newData: GameData) {
+    const Data = localStorage.getItem("gameResults");
+    let results: GameData[] = Data ? JSON.parse(Data) : [];
+
+    results.push(newData);
+
+    results.sort((a: GameData, b: GameData) => {
+      // 1) Lv 내림차순
+      if (b.level !== a.level) {
+        return b.level - a.level; // 큰 Lv가 앞에
+      }
+      // 2) score 오름차순
+      return a.score - b.score; // 작은 score가 앞에
+    });
+
+    results = results.slice(0, 3);
+    localStorage.setItem("gameResults", JSON.stringify(results));
+  }
+
   return (
     <main>
       <h2>다음 숫자 "{index + 1}"</h2>
@@ -51,13 +214,7 @@ const GameBoard: React.FC<GameBoardProps> = ({
                   <td key={colIndex}>
                     <GameButton
                       onClick={() => buttonHandler(idx)}
-                      className={
-                        clickCount[idx] === 1
-                          ? "clicked-once"
-                          : clickCount[idx] === 2
-                          ? "clicked-twice"
-                          : "clicked-none"
-                      }
+                      clickCount={clickCount[idx] || 0}
                     >
                       {numbers1[idx]}
                     </GameButton>
@@ -68,6 +225,14 @@ const GameBoard: React.FC<GameBoardProps> = ({
           ))}
         </tbody>
       </GameTable>
+      {isModalOpen && (
+        <Modal onClose={closeModal}>
+          <h2>clear!</h2>
+          <p>날짜 : {loadData().time}</p>
+          <p>Level : {loadData().level}</p>
+          <p>Time : {loadData().score}</p>
+        </Modal>
+      )}
     </main>
   );
 };
