+++ src/main.tsx
@@ -1,6 +1,5 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
 
 import AppWrapper from "./AppWrapper.tsx";
